requirements.txt
requests>=2.31
gTTS>=2.5
moviepy>=2.0
Pillow>=10.0
toml>=0.10
rich>=13.0
yt-dlp>=2024.0
playwright>=1.40
google-api-python-client>=2.100
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.1
